feat: add stylelint
