+++ clients/design-system/.stylelintrc.json
@@ -1,0 +1,16 @@
+{
+  "extends": "stylelint-config-standard-scss",
+  "rules": {
+    "alpha-value-notation": "number"
+  },
+  "overrides": [
+    {
+      "files": ["*.scss", "**/*.scss"],
+      "customSyntax": "postcss-scss",
+      "rules": {
+        "scss/no-global-function-names": null,
+        "selector-pseudo-class-no-unknown": null
+      }
+    }
+  ]
+}
