+++ clients/design-system/src/components/Alert/style.module.scss
@@ -1,27 +1,35 @@
 .alert {
- box-shadow: rgba(0, 0, 0, 0.12) 0px 4px 30px 0px;
+ box-shadow: rgb(0 0 0 / 0.12) 0 4px 30px 0;
  border-radius: 3px;
  border: none;
+
  &.alert-success {
   background-color: #f9fef9  !important;
+
   svg {
       color: #59ca97;
    }
   }
+
  &.alert-warning {
   background-color: #fefbf9 !important;
+
   svg {
     color: #f1963b;
    }
   }
+
  &.alert-info {
   background-color: #fafdff !important;
+
   svg {
     color: #1174ff;
   }
  }
+
  &.alert-danger {
   background-color: #fffafb !important;
+
   svg {
     color: #bf1722;
   }
