+++ clients/design-system/src/components/AppBar/style.module.scss
@@ -2,11 +2,13 @@
 
 @keyframes pulsate {
   0% {
-    box-shadow: inset 0 -3px 0 rgba(0, 0, 0, 0);
+    box-shadow: inset 0 -3px 0 rgb(0 0 0 / 0);
   }
+
   50% {
-    box-shadow: inset 0 -3px 0 rgba(0, 0, 0, 0.2);
+    box-shadow: inset 0 -3px 0 rgb(0 0 0 / 0.2);
   }
+
   100% {
     box-shadow: inset 0 -3px 0 $color-blue;;
   }
@@ -14,9 +16,11 @@
 
 .app-bar {
   background-color: $color-white;
+
   .navbar-menu {
-    box-shadow: rgba(0, 0, 0, 0.12) 0px 4px 30px 0px
+    box-shadow: rgb(0 0 0 / 0.12) 0 4px 30px 0
   }
+
   .navbar-brand {
     font-weight: bold;
     font-size: 2rem !important;
@@ -27,30 +31,34 @@
 
 .app-bar-light {
   padding: 0;
-  background-color: rgb(255, 253, 251);
-  box-shadow: rgba(0, 0, 0, 0.12) 0px 0px 2px 0px, rgba(0, 0, 0, 0.14) 0px 1.008px 2px 0px;
+  background-color: rgb(255 253 251);
+  box-shadow: rgb(0 0 0 / 0.12) 0 0 2px 0, rgb(0 0 0 / 0.14) 0 1.008px 2px 0;
 
   ul {
     padding: 0;
     margin: 0;
+
     li {
       margin-right: 2rem;
       padding: 0;
+
       a {
         color: $color-blue;
         display: inline-block;
         text-decoration: none;
         padding-top: 1rem;
         padding-bottom: 1rem;
+
         &:hover {
           color: $color-blue;
           animation: pulsate 0.5s ease-in-out forwards;
           transition: border-color 0.5s ease;
-          box-shadow: inset 0 -4px 0 rgba(0, 123, 255, 0.5);
+          box-shadow: inset 0 -4px 0 rgb(0 123 255 / 0.5);
         }
+
         &.active {
           color: $color-blue;
-          box-shadow: inset 0 -4px 0 rgba(0, 123, 255, 0.5);
+          box-shadow: inset 0 -4px 0 rgb(0 123 255 / 0.5);
         }
       }
     }
