+++ clients/design-system/src/components/BlurbPair/BlurbIconColors.enum.ts
@@ -1,0 +1,6 @@
+export enum BlurbIconColors {
+  BLUE = 'blue',
+  ORANGE = 'orange',
+  GREEN = 'green',
+  PURPLE = 'purple',
+}
