+++ clients/design-system/src/components/BlurbPair/style.module.scss
@@ -9,18 +9,22 @@
       }
     }
   }
+
   .blurb-pair-left,
   .blurb-pair-right {
     .icon {
       &.icon-color-blue {
         color: $color-blue !important;
       }
+
       &.icon-color-orange {
         color: $color-orange !important;
       }
+
       &.icon-color-green {
         color: $color-light-green !important;
       }
+
       &.icon-color-purple {
         color: $color-dark-purple !important;
       }
