+++ clients/design-system/src/components/HeroBanner/style.module.scss
@@ -23,19 +23,22 @@ $hero-banner-height: 300px;
       left: 0;
       width: 100%;
       height: 100%;
-      background: linear-gradient(90deg, rgba(128, 80, 237,0.9) 0%, rgba(128, 80, 237,0.9) 28%, rgba(0, 141, 184,0.5) 100%);
+      background: linear-gradient(90deg, rgb(128 80 237 / 0.9) 0%, rgb(128 80 237 / 0.9) 28%, rgb(0 141 184 / 0.5) 100%);
       clip-path: polygon(0 0, 100% 0, 100% 100%, 100% 100%, 0% 50%);
       z-index: -1;
     }
 
     .hero-text-wrapper {
       background-color: $color-dark-blue;
+
       h1 {
         font-weight: bold;
       }
+
       .lead {
         color: $color-gray;
       }
+
       p {
         color: $color-gray;
       }
