+++ clients/design-system/src/styles/utils/_bg-color.scss
@@ -1,26 +1,26 @@
 @import './color-pallette';
 
-.bgOffWhite {
+.bg-off-white {
   background-color: $color-off-white !important;
 }
 
-.bgBrandDark {
+.bg-brand-dark {
   background-color: $color-berkeley-blue !important;
 }
 
-.bgBrandLight {
+.bg-brand-light {
   background-color: $color-gold !important;
 }
 
-.bgBrandPrimary {
+.bg-brand-primary {
   background-color: $color-murrey !important;
 }
 
-.bgBrandSecondary {
+.bg-brand-secondary {
   background-color: $color-sgbus-green !important;
 }
 
-.bgBrandTertiary {
+.bg-brand-tertiary {
   background-color: $color-sky-blue !important;
 }
