+++ clients/design-system/src/styles/utils/_color-pallette.scss
@@ -2,21 +2,17 @@
 $color-gold: #ffd700;
 $color-sgbus-green: #71cd2f;
 $color-murrey: #b3105c;
-
 $color-white: #FFF;
 $color-off-white: #f5f5fa; // #f4fafd; #f1f4f9;
 $color-cta-white: #e4f2ff;
 $color-light-gray: #dddde5;
-$color-gray:  #9697a6;
+$color-gray: #9697a6;
 $color-off-black: #1e1e1f;
 $color-black: #000;
-
 $color-sky-blue: #87ceeb;
-$color-blue:  #0090BA;
+$color-blue: #0090BA;
 $color-berkeley-blue: #094581;
 $color-dark-blue: #08062a;
-
-
 $color-orange: #fdb913;
 $color-light-green: #00eb80;
 $color-dark-purple: #8050ed;
