+++ common/config/rush/repo-state.json
@@ -1,5 +1,5 @@
 // DO NOT MODIFY THIS FILE MANUALLY BUT DO COMMIT IT. It is generated and used by Rush.
 {
-  "pnpmShrinkwrapHash": "ad0a9f19815b6f96bdcdf4ecfa8bb99eb710d768",
+  "pnpmShrinkwrapHash": "8d1480817cfd5089ab844be5e8e5eae49c14098b",
   "preferredVersionsHash": "a48003cf229dd47d077bcf6301ac15a6f90e1c34"
 }
